code/leetcode/simple-and-midium/t6_test.go
package simple_and_midium

import "testing"

func Test_convert(t *testing.T) {
	type args struct {
		s       string
		numRows int
	}
	tests := []struct {
		name string
		args args
		want string
	}{
		{
			args: args{
				s: "A",
				numRows: 1,
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := convert(tt.args.s, tt.args.numRows); got != tt.want {
				t.Errorf("convert() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_convert2(t *testing.T) {
	type args struct {
		s       string
		numRows int
	}
	tests := []struct {
		name string
		args args
		want string
	}{
		{
			args: args{
				s: "PAYPALISHIRING",
				numRows: 4,
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := convert2(tt.args.s, tt.args.numRows); got != tt.want {
				t.Errorf("convert2() = %v, want %v", got, tt.want)
			}
		})
	}
}